+++ src/main.rs
@@ -7,6 +7,7 @@ mod card;
 mod menu;
 mod read;
 mod run;
+mod settings;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let args = std::env::args().collect::<Vec<String>>();
@@ -21,6 +22,23 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             )
         }
     };
+
+    let settings_path = match option_env!("SETTINGS_PATH") {
+        Some(path) => path,
+        None => {
+            return Err(
+                "This program needs to be compiled with the $SETTINGS_PATH env variable set".into(),
+            )
+        }
+    };
+    let settings = match settings::Settings::load(settings_path) {
+        Ok(settings) => settings,
+        Err(e) => {
+            error!("Failed to load settings: {}", e);
+            return Ok(());
+        }
+    };
+    debug!("Settings: {:?}", settings);
 
     let config = match config::RunMode::from_args(&args) {
         Ok(config) => config,